+++ dart/test/sentry_event_test.dart
@@ -285,5 +285,27 @@ void main() {
       expect(eventMap['request'], isNull);
       expect(eventMap['debug_meta'], isNull);
     });
+
+    test(
+        'throwable and originThrowable should return origin throwable if no mechanism',
+        () {
+      final error = StateError('test-error');
+      final event = SentryEvent(throwable: error);
+
+      expect(event.throwable, error);
+      expect(event.originThrowable, error);
+    });
+
+    test(
+        'originThrowable getter should return origin throwable if theres a mechanism',
+        () {
+      final error = StateError('test-error');
+      final mechanism = Mechanism(type: 'FlutterError', handled: true);
+      final throwableMechanism = ThrowableMechanism(mechanism, error);
+      final event = SentryEvent(throwable: throwableMechanism);
+
+      expect(event.throwable, throwableMechanism);
+      expect(event.originThrowable, error);
+    });
   });
 }